LayoutTests/http/tests/resourceLoadStatistics/strip-referrer-to-origin-for-third-party-requests-expected.txt

Tests that only the origin is sent as referrer for third-party requests.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS referrer is "http://127.0.0.1:8000/resourceLoadStatistics/strip-referrer-to-origin-for-third-party-requests.html"
PASS referrer is "http://127.0.0.1:8000/"
PASS successfullyParsed is true

TEST COMPLETE


--------
Frame: '<!--frame1-->'
--------
http://127.0.0.1:8000/
